+++ components/form/context.tsx
@@ -16,7 +16,6 @@ export interface FormContextProps {
   labelAlign?: FormLabelAlign;
   labelCol?: ColProps;
   wrapperCol?: ColProps;
-  itemReferable?: boolean;
   itemRef: (name: (string | number)[]) => (node: React.ReactElement) => void;
 }
