+++ components/form/index.en-US.md
@@ -25,7 +25,6 @@ High performance Form component with data scope management. Including data colle
 | form | Form control instance created by `Form.useForm()`. Automatically created when not provided | [FormInstance](#FormInstance) | - |  |
 | hideRequiredMark | Hide required mark for all form items | boolean | false |  |
 | initialValues | Set value by Form initialization or reset | object | - |  |
-| itemReferable | Field instance can be get by `getFieldInstance` | boolean | false | 4.4.0 |
 | labelAlign | text align of label of all items | `left` \| `right` | `right` |  |
 | labelCol | label layout, like `<Col>` component. Set `span` `offset` value like `{span: 3, offset: 12}` or `sm: {span: 3, offset: 12}` | [object](/components/grid/#Col) | - |  |
 | layout | Form layout | `horizontal` \| `vertical` \| `inline` | `horizontal` |  |
@@ -87,7 +86,6 @@ Form field component for data bidirectional binding, validation, layout, and so 
 | labelCol | The layout of label. You can set `span` `offset` to something like `{span: 3, offset: 12}` or `sm: {span: 3, offset: 12}` same as with `<Col>`. You can set `labelCol` on Form. If both exists, use Item first | [object](/components/grid/#Col) | - |  |
 | name | Field name, support array | [NamePath](#NamePath) | - |  |
 | normalize | Normalize value from component value before passing to Form instance | (value, prevValue, prevValues) => any | - |  |
-| referable | Field instance can be get by `getFieldInstance` | boolean | false | 4.4.0 |
 | required | Display required style. It will be generated by the validation rule | boolean | false |  |
 | rules | Rules for field validation. Click [here](#components-form-demo-basic) to see an example | [Rule](#Rule)[] | - |  |
 | shouldUpdate | Custom field update logic. See [below](#shouldUpdate) | boolean \| (prevValue, curValue) => boolean | false |  |
@@ -189,7 +187,7 @@ Provide linkage between forms. If a sub form with `name` prop update, it will au
 
 | Name | Description | Type |
 | --- | --- | --- |
-| getFieldInstance | Get field instance which config `referable` | (name: [NamePath](#NamePath)) => any | 4.4.0 |
+| getFieldInstance | Get field instance | (name: [NamePath](#NamePath)) => any | 4.4.0 |
 | getFieldValue | Get the value by the field name | (name: [NamePath](#NamePath)) => any |
 | getFieldsValue | Get values by a set of field names. Return according to the corresponding structure | (nameList?: [NamePath](#NamePath)[], filterFunc?: (meta: { touched: boolean, validating: boolean }) => boolean) => any |
 | getFieldError | Get the error messages by the field name | (name: [NamePath](#NamePath)) => string[] |
