+++ components/form/index.zh-CN.md
@@ -26,7 +26,6 @@ cover: https://gw.alipayobjects.com/zos/antfincdn/YeW5m96xHG/Form.svg
 | form | 经 `Form.useForm()` 创建的 form 控制实例，不提供时会自动创建 | [FormInstance](#FormInstance) | - |  |
 | hideRequiredMark | 隐藏所有表单项的必选标记 | boolean | false |  |
 | initialValues | 表单默认值，只有初始化以及重置时生效 | object | - |  |
-| itemReferable | 字段元素是否可被 `getFieldInstance` 获取 | boolean | false | 4.4.0 |
 | labelAlign | label 标签的文本对齐方式 | `left` \| `right` | `right` |  |
 | labelCol | label 标签布局，同 `<Col>` 组件，设置 `span` `offset` 值，如 `{span: 3, offset: 12}` 或 `sm: {span: 3, offset: 12}` | [object](/components/grid/#Col) | - |  |
 | layout | 表单布局 | `horizontal` \| `vertical` \| `inline` | `horizontal` |  |
@@ -88,7 +87,6 @@ const validateMessages = {
 | labelCol | `label` 标签布局，同 `<Col>` 组件，设置 `span` `offset` 值，如 `{span: 3, offset: 12}` 或 `sm: {span: 3, offset: 12}`。你可以通过 Form 的 `labelCol` 进行统一设置。当和 Form 同时设置时，以 Item 为准 | [object](/components/grid/#Col) | - |  |
 | name | 字段名，支持数组 | [NamePath](#NamePath) | - |  |
 | normalize | 组件获取值后进行转换，再放入 Form 中 | (value, prevValue, prevValues) => any | - |  |
-| referable | 字段元素是否可被 `getFieldInstance` 获取 | boolean | false | 4.4.0 |
 | required | 必填样式设置。如不设置，则会根据校验规则自动生成 | boolean | false |  |
 | rules | 校验规则，设置字段的校验逻辑。点击[此处](#components-form-demo-basic)查看示例 | [Rule](#Rule)[] | - |  |
 | shouldUpdate | 自定义字段更新逻辑，说明[见下](#shouldUpdate) | boolean \| (prevValue, curValue) => boolean | false |  |
@@ -190,7 +188,7 @@ Form 通过增量更新方式，只更新被修改的字段相关组件以达到
 
 | 名称 | 说明 | 类型 | 版本 |
 | --- | --- | --- | --- |
-| getFieldInstance | 获取对应字段示例，需要字段开启 `referable` | (name: [NamePath](#NamePath)) => any | 4.4.0 |
+| getFieldInstance | 获取对应字段示例 | (name: [NamePath](#NamePath)) => any | 4.4.0 |
 | getFieldValue | 获取对应字段名的值 | (name: [NamePath](#NamePath)) => any |  |
 | getFieldsValue | 获取一组字段名对应的值，会按照对应结构返回 | (nameList?: [NamePath](#NamePath)[], filterFunc?: (meta: { touched: boolean, validating: boolean }) => boolean) => any |  |
 | getFieldError | 获取对应字段名的错误信息 | (name: [NamePath](#NamePath)) => string[] |  |
