+++ components/form/interface.ts
@@ -1,3 +1,3 @@
 export { Options as ScrollOptions } from 'scroll-into-view-if-needed';
 export type FormLabelAlign = 'left' | 'right';
-export { Store, StoreValue } from 'rc-field-form/lib/interface';
+export { Store, StoreValue, NamePath, InternalNamePath } from 'rc-field-form/lib/interface';
